change type of Cr

+++ model/glomerular_infiltration.go
@@ -15,6 +15,6 @@ type (
 
 	GlomerularInfilLog struct {
 		Date time.Time `json:"date,omitempty" bson:"date"`
-		Cr   int       `json:"cr" bson:"cr"`
+		Cr   float64   `json:"cr" bson:"cr"`
 	}
 )
